two fixes that help autoresize work in IE6 and when using style sheets.
First couple chanes fix bugs in IE6.
doDelayedResize is to fix a bug with slow loading css files. If a css file took to long to load under the original method, it would autoresize incorrectly. This fixes the problem such that it waits atleast 200ms after the last style sheet loads before doing the final resize.