Fixed bug #5881.
Tested on iOs WebKit: If the Editor's origin document element is different than the actual contentEditable's document then it needs a focus call on the contentEditable's window object to have the correct focus again.
Without this the contentEditable element does not have the proper focus and therefore keyboard input has no effect to the actual content.