Create plugin.selected.js
Hi I have modified the code plugin to allow a selection of code to me edited if an editor selection is active. Otherwise the default will be to load the editor code as per original implementation.

I am not an experienced TinyMCE plugin developer, so this code may need refinement. This is just an idea as I need this form of functionality to help my staff edit large chunks of HTML.

Richard Hetherington