Update image plugin - IE8 bug
In IE8 the done event was firing before it was attached to the dom.  This was throwing a js error when trying to access the parentNode.removeChild method.  Reordering the events to be attached after the image is appended to the dom fixes this problem, since the image will now be in the dom when the done method is triggered.