Fix indentation feature for forced_root_block==false.
If forced_root_block==false and the text to indent is not inside a block element a block element will be added to apply the indentation.
With this fix, this also works for textnodes which are direct children of the editor.bodyElement.

The new implementation fixes some problems of the current implementation:
- Using selection.getRng().commonAncestorNode instead of selection.getNode(), as selection.getNode() does not return text nodes.
- The search of a parent block element is now limited to parents inside the editor.bodyElement