Add CloseQuery event to FloatPanel
This event enables code, which creates a custom dialog, to prevent the user from closing the dialog window by calling evt.preventDefault().
This can be useful to prohibit closing the dialog during critical sections of the process or to ask the user if they really want to close the window without submitting.