Prevent handled tab events from propagating outside the editor.
Allow 'tab' to escape from non-modal elements. Otherwise, the
containing application might assume that it should handle it.

Currently, TinyMCE always propagates tab key events. Unfortunately,
containing applications have little way of knowing whether this is a
tab that they should intercept -- such as in the toolbar -- or a tab
they should most definitely ignore -- such as in menus and dialogs.

We handle this by only letting tab keys propagate when unhandled; in
order to prevent toolbars from becoming tab sinks, we also add a
notion of controls that wrap focus. The latter currently applies to
menus and windows.