should not keep its own "domReady" state
It should always check the current DOM state to see if it is ready or
not, and not keep a internal DOM state tracker.

Reason: If any other lib is also listening to a DOM Ready call and get
the call before TinyMCE, and that JS code initialize TinyMCE, TinyMCE
code will delay the its initialization (although the DOM is already
ready) until the initial JS code executions finishes.

TinyMCE should not store an internal DOM state internally, but should
always check directly in the DOM to verify if the DOM is ready or not.

less version should be less then 2.0.0

phantomjs install should not use -g params