Add global hook into XRH setup of each request
There are times where the data sent in the XHR request needs to be modified.

An example of where this is needed can be seen in [this pull request](https://github.com/tinymce/tinymce/pull/498).  In this example, the plugin does't currently allow the URL to be modified at the time an XHR request is made.  Having a global hook, is a way to access the request even if the plugin hasn't provided for it.

This commit adds a hook before the XHR object is initialised, allowing for the settings to be modified at the time of the request.

```
tinymce.util.XHR.on('beforeInitialize', function(settings) {
    if (settings.data) settings.data = settings.data + '&';
    settings.data = settings.data + "foo=bar";
});
```