Fix missing call of the custom url converter in the image plugin
Bugfix for the image plugin. If a custom URL converter is configured (configuration option "convert_urls"), it will be called after changes of the source field in the "insert/edit image" dialog. 
Unfortunately the url converter is not called if the user clicks the "Ok" button directly (form submission).
The problem occured with the Internet Explorer 9.



If the configuration has defined an URL converter . It will be called if you make changes in the src Field

Bugfix for the image plugin. Also call the convertURL