Ensure that the offscreen selection remains offscreen.
When selecting a non-editable block containing a table with styles specifying "width: 100%; table-layout:fixed;", the table extends into the visible area on Internet Explorer 11. (Firefox and Chrome appear to limit the width of elements to 1 million pixels, which remains off-screen)

Correcting the height and width specification of the mce-offscreen-selection block ensures that the non-editable table remains offscreen.