Prevent a situation in which for example selecting all and cutting in a WebKit browser will remove typing focus. Whenever a caret node is sought, it should be a pre-requisite that the node in question is still part of the editor's DOM. In some circumstances where the start or end block for a deletion is completely removed, it is still considered as a viable block for restoring caret.
I was not able to come up with a unit test to cover this change because I think the "unfocused" state reached in a side-effect of setting the selection to a DOM node that has been removed from the tree, but the selection "fixes itself up" at least as far as the DOM is concerned. So although from a practical point of view the editor has lost focus, testing the selection with getRng() shows that it has a reasonable selection as 0-index relative to the editor body.