docs: Fix misleading info in SelectBox documentation
The SelectBox documentation stated that in order to provide
values to serve as options when the HTML component would be rendered,
a `values` array should be passed, and its items would be used
as option values.
In the code, in `SelectBox.js:55` the property being checked is
`self.options` and there is no occurrence of `self.values` which
probably means that the documentation part was an unfortunate
copy+paste.

Signed-off-by: Adrian Oprea <adrian@codesi.nz>