Added tests for using the base i18n language if a variant is in use.
If a variant of a base language is in use (e.g. "fr_FR") and does not contain the required translation, the base language (e.g. "fr") should also be consulted. In these tests, "only_base" should be translated (from the base language), "not_defined_anywhere" should be unchanged, and the other translations should come from the variant.