Consult the base language for a translation if it is missing in the variant.
If a variant of a base language is in use (e.g. "fr_FR") and does not contain the required translation, the base language (e.g. "fr") should also be consulted. getTranslated() first consults the currently selected language for a translation; if none is found, and the currently selected language is a variant, it also consults the base language.