plugins/autosave - Add 'setPrefix' and 'getPrefix'
Motivation: For a web2py + Vue.js project I had routes after the hash in the URL and so wasn't able to use the autosave plugin because the LocalStorage prefix (key) was the same for different pages. Being able to set and get the prefix fixed this issue. Being able to include document.location.hash in the prefix would also have fixed this issue (so I've added that ability here as well).

So, for example, I'd go to www.mywebsite.com/web2py_app_name/controller_name/function_name#/customers/view/73, and that would have the same prefix as the entity with an ID of 74.