modified the plugin charmap in order to use combined characters/signs instead of only one.
The plugin accepts now either a hexcode (which has been standard so far) or an array consisting of hexcodes as first argument. Using the array allows us to combine diacritics, etc.
A third argument was introduced as well carrying a description or explanation for the particular character that will be inserted.