TINY-1615: Restrict Arr functions input to ReadonlyArray<T>
None of the Arr functions mutate the passed array and this should be
represented in the typesystem by using ReadonlyArray<T>. This also has
the advantage that arrays frozen by Object.freeze can be passed without
having to first convert with Arr.from.

Also added tests to ensure that the Arr functions can be called with
both mutable and frozen arrays.