Fixed an issue that was causing find and replace to replace the wrong content in certain circumstances
If a line contained new chars (eg <br />), then if a search was done for a blank space, the find and replace would incorrectly match to those new line chars. That would then cause the find logic to remove the chars after the new lines from view when doing a find and then add additional chars when replacing.