- Creates formula plugin.
This plugin is based on the charmap plugin but focused for mathematical formulas where the user can select variables from a list of variables.

The variables are paramatized. Variables can have a description of what it references to and the contracted form which will be displayed in the mathematical equation.