Remove Cell that was of no use.
Storage.startStoreDraft is only called from Buttons.register, which is only called from a closure in Plugin. The closure in Plugin creates a new Cell(false) as a local variable to pass in. As such, startStoreDraft is always called with a Cell(false). It sees that it is false, makes the setInterval call, sets the Cell to true... then the Cell falls out of scope, to  be garbage-collected.