TINY-6412: Refactor ZoneWalker part 6
Here we delete the ZoneWalkerState ADT / combine the analyse + process functions.
As the traversal logic got pulled up to the walk function in the last two
commits I noticed that the true "state" we were really passing around was the
Traverse object. The ZoneWalkerState was only ever created to get passed from
analyse to process. That's been taken out now.