Fix disappearing words during Word list paste cleanup
Fixes #2810, Fixes #3017, Fixes #3480, Fixes #5474

This commit includes fixes to multiple subtle errors in MS Word list
cleanups. The bugs caused any characters ending with a dot to disappear
from the lists under certain circumstances:

Font/style -changes cause multiple text nodes to appear as siblings in
the AST.

Function filterStyles marked only the first node with _listIgnore and
thus missed the nodes with roman numerals. This was fixed.

The removeIgnoredNodes-function used node.remove() to delete nodes
from the AST. The method sets node.next to null and thus recursion
did not proceed further. Sibling text nodes with _listIgnore would
not be deleted. This was fixed by avoiding node.remove().

The trimListStart -function did not trim the beginning of the
paragraph. It trimmed all text nodes instead and removed the ones with
words ending with dot. This was fixed by mering the paragraph
text nodes before trimming the first text node.

Regex for cleaning up the beginning is now safe. It no longer removes
any letters (with \w). Instead it only looks for certain bullets. One
regex is enough to filter those as the text nodes have been merged
before matching.

Added a couple of test cases as a proof that the fixes work correctly.