Treat the default/undefined `schema` as 'html5' as specified by docs.
The documentation claims that [the default schema is 'html5'](https://www.tiny.cloud/docs/configure/content-filtering/#schema) but nowhere is this default actually specified so the default value ends up being `undefined`. This is fine for code like `schema !== 'html4'` but doesn't work for `schema !== 'html5'` since it will evaluate to true when the schema is undefined.

The impact of this problem is that `schema: 'html5'` actually differs from omitting the schema. Only an explicit 'html5' actually disables the conversion from <b> => <strong> and <i> => <em>.

This change defaults the schema to 'html5' when `undefined`.