Fix the color contrast issues for axe a11y audit
Elements with the following CSS classes fire the "Text elements must
have sufficient color contrast against the background" a11y violation in
the axe DevTools:

- [`tox-mbtn__select-label`](https://axe.deque.com/issues/7babe2d7-8d6f-4f82-a3f9-9f38e3d6ab09)
- [`tox-tbtn__select-label`](https://axe.deque.com/issues/a190e94d-b8fe-41f5-aa4c-ad14ef6fef4d)

https://dequeuniversity.com/rules/axe/4.3/color-contrast?application=AxeChrome

As far as I understand the only purpose of `@toolbar-background` and
`@menubar-background` is to dynamically generate a transparent SVG
background with a 1px bottom line.

So I'd suggest achieving absolutely the same result by setting the
`border-bottom` property and as a benefit improve accessibility.