TINY-8952: Adding support for Searchable Menus. This included a lot of little changes
alloy
====

TieredMenu
----------

* changed highlightImmediately flag in TieredMenu to the more granular HighlightOnOpen.
We needed to differentiate between highlighting nothing, highlighting just the menu, and
highlighting the menu and the item.

* started to deprecate onHighlight. This function was confusing because its first
argument was the menu inside the TieredMenu, not the menu itself. This was because
it was just passing it on to its individual menu configurations. So instead, we have
created new handlers (onHighlightItem and onDehighlightItem) which will pass
all of the TieredMenu component, the Menu component, and the highlighted item Component. To
make them work, we've introduced two internal events to alloy that bubble from the
Menu highlight handlers to the TieredMenu

* introduced slightly better handling for fakeFocus in TieredMenus setActiveMenuAndItem.
The support for fakeFocus is still spotty in situations where the real focus is inside
the menu also, because using event.target can lead to the wrong information
(see in-code comment). This needs to be resolved properly in a separate JIRA.

Menu
----

* introduced onDehighlight. This enables us to make more complex ARIA configurations without
changing alloy (e.g. aria-selected for the SearchableMenu in TinyMCE)

Typeahead
---------

* introduced lazyTypeaheadComp to Typeahead. This resolved an issue where the communication
between Typeahead's Menu and the Typeahead Input relied on them being in the same mothership,
which isn't necessarily true. Now, When the Typeahead is attached, it stores a reference
to itself in the detail, which its parts have access to.

* tweaked interactions between selectsOver, highlighting, previewing etc. Because selectsOver
is not used in production code currently, this hasn't been sufficiently tested. The main gist
of the changes is ensuring that previewing mode is updating correctly, because when the Typeahead
is in prevewing mode, pressing <enter> behaves differently than when it isn't previewing.

Dropdown
--------

* introduce a new API called refetch. The Refetch API will still update the dropdown, even
if it is already open, and will return a Future, so that the calling site can be do something
once it has completed. This enables us to update the Dropdowns without closing the menu first.

Coupling
--------

* adding an API to Coupling (getExistingCoupled) that allows determining if the Coupling component
has been created, without creating it if it doesn't exist. This was used for refetching in Dropdowns.

AlloyTriggers
-------------

* introduced a new convenience function for dispatching an event with a *changed* target. This is
useful for scenarios where something with fakeFocus needs to be given an event that occurred on
something with real focus (e.g SearchableMenu). The function is retargetAndDispatchWith

bridge
======

* added a field to MenuButtons for whether or not its attached Menu was searchable.

* added fetchContext to MenuButtons fetch function, which currently exposes the search pattern
that was typed in the SearchableMenu. This allows the integrator to filter the responses based
on that SearchPattern.

* TODO: add a field to MenuButtons for setting the placeholder text of the MenuButton search

silver
======

* introduced SearchableMenus as a concept. This involved
  * creating menuLayouts for searchable menus with the field, and without it including any ARIA
interactions. This was separated from Toolbar.PresetTypes, so that it didn't impact customer API
  * reworking of NestedMenus.build API to allow searchable flags
  * passing the search pattern back to MenuButton's fetch call
  * keeping focus in the SearchableMenu field, while still allowing navigation of the menu

The ARIA interactions are documented in TINY-8952. A SearchableMenu works with keeping the focus
inside the search input field, and using FakeFocus to navigate between the items. It does not
enforce any collapsing of results when a search pattern is present, though that is recommended.
The search field, itself, should not be within the FakeFocus navigation's cycle.

Because left, right, and space have value inside a field, these keystroke are not processed by the
menu. The equivalent expand/collapse actions on items can only be triggered by Enter/Escape, because
these keystrokes have no useful purpose in the search field.