TINY-8952: Adding support for Searchable Menus. This included a lot of little changes (#8058)
* TINY-8952: Adding support for Searchable Menus. This included a lot of little changes

alloy
====

TieredMenu
----------

* changed highlightImmediately flag in TieredMenu to the more granular HighlightOnOpen.
We needed to differentiate between highlighting nothing, highlighting just the menu, and
highlighting the menu and the item.

* started to deprecate onHighlight. This function was confusing because its first
argument was the menu inside the TieredMenu, not the menu itself. This was because
it was just passing it on to its individual menu configurations. So instead, we have
created new handlers (onHighlightItem and onDehighlightItem) which will pass
all of the TieredMenu component, the Menu component, and the highlighted item Component. To
make them work, we've introduced two internal events to alloy that bubble from the
Menu highlight handlers to the TieredMenu

* introduced slightly better handling for fakeFocus in TieredMenus setActiveMenuAndItem.
The support for fakeFocus is still spotty in situations where the real focus is inside
the menu also, because using event.target can lead to the wrong information
(see in-code comment). This needs to be resolved properly in a separate JIRA.

Menu
----

* introduced onDehighlight. This enables us to make more complex ARIA configurations without
changing alloy (e.g. aria-selected for the SearchableMenu in TinyMCE)

Typeahead
---------

* introduced lazyTypeaheadComp to Typeahead. This resolved an issue where the communication
between Typeahead's Menu and the Typeahead Input relied on them being in the same mothership,
which isn't necessarily true. Now, When the Typeahead is attached, it stores a reference
to itself in the detail, which its parts have access to.

* tweaked interactions between selectsOver, highlighting, previewing etc. Because selectsOver
is not used in production code currently, this hasn't been sufficiently tested. The main gist
of the changes is ensuring that previewing mode is updating correctly, because when the Typeahead
is in prevewing mode, pressing <enter> behaves differently than when it isn't previewing.

Dropdown
--------

* introduce a new API called refetch. The Refetch API will still update the dropdown, even
if it is already open, and will return a Future, so that the calling site can be do something
once it has completed. This enables us to update the Dropdowns without closing the menu first.

Coupling
--------

* adding an API to Coupling (getExistingCoupled) that allows determining if the Coupling component
has been created, without creating it if it doesn't exist. This was used for refetching in Dropdowns.

AlloyTriggers
-------------

* introduced a new convenience function for dispatching an event with a *changed* target. This is
useful for scenarios where something with fakeFocus needs to be given an event that occurred on
something with real focus (e.g SearchableMenu). The function is retargetAndDispatchWith

bridge
======

* added a field to MenuButtons for whether or not its attached Menu was searchable.

* added fetchContext to MenuButtons fetch function, which currently exposes the search pattern
that was typed in the SearchableMenu. This allows the integrator to filter the responses based
on that SearchPattern.

* TODO: add a field to MenuButtons for setting the placeholder text of the MenuButton search

silver
======

* introduced SearchableMenus as a concept. This involved
  * creating menuLayouts for searchable menus with the field, and without it including any ARIA
interactions. This was separated from Toolbar.PresetTypes, so that it didn't impact customer API
  * reworking of NestedMenus.build API to allow searchable flags
  * passing the search pattern back to MenuButton's fetch call
  * keeping focus in the SearchableMenu field, while still allowing navigation of the menu

The ARIA interactions are documented in TINY-8952. A SearchableMenu works with keeping the focus
inside the search input field, and using FakeFocus to navigate between the items. It does not
enforce any collapsing of results when a search pattern is present, though that is recommended.
The search field, itself, should not be within the FakeFocus navigation's cycle.

Because left, right, and space have value inside a field, these keystroke are not processed by the
menu. The equivalent expand/collapse actions on items can only be triggered by Enter/Escape, because
these keystrokes have no useful purpose in the search field.

* TINY-8952: Moving Coupling tests into coupling directory

* TINY-8952: Adding tests for getExistingCoupled, and cleaning up existing behaviour

* TINY-8952: AlloyTriggers retargeting test

* TINY-8952: Testing Dropdown.refetch

* TINY-8952: Menu tests for onHighlight and onDehighlight

* TINY-8952: Adding test for fakeFocus for TieredMenus

* TINY-8952: Adding tests for onHighlight, on HighlightItem and onDehighlightItem - but onDehighlight is firing too often

* TINY-8952: Fixing dehighlightExcept in HighlightingApis.

This fix allowed us to remove the redundant onDehighlight, onHighlight expectations in TieredMenuTest

* TINY-8952: HighlightOnOpen tests

* TINY-8952: Fixing populateFromBrowse and adding more Typeahead tests for different model configurations

* TINY-8952: Typeaheads when input and menus are in different motherships

* TINY-8952: SearchableMenu testing

* aria and DOM structures

No webdriver tests yet.

* TINY-8952: Webdriver tests. Highlighting problem with Esc when the multiple menus are visible

the issue is that Escape on input[type=search] by default clears the input, which triggers an input event, which refetches, which makes one menu, and then the Esc keyup fires, and that closes that one menu.

* TINY-8952: Fixing Escape key closing all menus when a search pattern is present

* this was caused by input[type=search] handling Esc natively, and clearing the Input field, which triggered an Input event, which refetched, resulting in just one menu. Then the keyup of Esc fired and closed that one menu.

* TINY-8952: Fixing export issue for agar

* TINY-8952: eslint issue

* TINY-8952: Customisable filter placeholders for SearchableMenus

* updates to boulder to allow converging of different schema into one value
* reworking of menu search types and layout types
* passing through placeholder to search field in menus

* TINY-8952: Do not have a default Placeholder for Searchable Menus

* TINY-8952: Adding JIRA reference for fake focus handling

* TINY-8952: New search field for MenuButtons changelog entry

* TINY-8952: Apply suggestions from code review

Co-authored-by: Lee Newson <lee.newson@tiny.cloud>
Co-authored-by: Andrew Herron <thespyder@programmer.net>

* TINY-8952: More PR fixes

* TINY-8952: Renaming test

* TINY-8952: Clearing reference on detached

* TINY-8952: Update modules/alloy/CHANGELOG.md

Co-authored-by: Lee Newson <lee.newson@tiny.cloud>

* TINY-8952: Apply suggestions from code review

Co-authored-by: Lee Newson <lee.newson@tiny.cloud>

* TINY-8952: More PR fixes

* TINY-8952: Reverting MenuSpec rename

* TINY-8952: More PR fixes

Co-authored-by: Lee Newson <lee.newson@tiny.cloud>
Co-authored-by: Andrew Herron <thespyder@programmer.net>