TINY-9829: Add paste event to event order assertion in InternalClipboardTest
- Change init_instance_callback option to setup: The paste overrides (which fire the other events) happen in a `paste` event handler.
Having the test handlers in `init_instance_callback` will make them run after the `paste` handler with overrides. This would make it
look like the `paste` event is fired after all the other events but in reality it is fired first. This is why `setup` is used to make
the test handlers run before the any handlers in the paste implementation.