TINY-9829: Dispatch "insertFromPaste" `beforeinput` event with `dataTransfer` object on paste (#8747)
* TINY-9829: Dispatch beforeinput event with dataTransfer object before dispatching input event on paste

* TINY-9829: Update changelog entry

* TINY-9829: Use DataTransferContent module to handle setting html data

* TINY-9829: Remove redundant async await statement from pWaitForAndAssertNoEvents in InternalClipboardTest

* TINY-9829: Add assertion for paste event ordering in InternalClipboardTest

* TINY-9829: Add paste event to event order assertion in InternalClipboardTest

- Change init_instance_callback option to setup: The paste overrides (which fire the other events) happen in a `paste` event handler.
Having the test handlers in `init_instance_callback` will make them run after the `paste` handler with overrides. This would make it
look like the `paste` event is fired after all the other events but in reality it is fired first. This is why `setup` is used to make
the test handlers run before the any handlers in the paste implementation.

* TINY-9829: Use generics for setEventSingletonAndAddType in InternalClipboardTest