Trim titles and values in basic settings (#9598)
The `buildBasicSettingsDataset` call is used for:
- `block_formats`
- `font_family_formats`
- `font_size_formats`

The documented values for both `block_formats` and
`font_family_formats` include a space between the `;` and the title, for
example the default value in `block_formats` is:

```
Paragraph=p; Heading 1=h1; Heading 2=h2; Heading 3=h3; Heading 4=h4; Heading 5=h5; Heading 6=h6;
```

However the `split()` method in `SelectDatasets.ts` splits on the
delimiter, leading to values such as ` Heading 1=h1` which are then
passed into the `process()` method which just split them on `=` leading
to a title of ` Heading 1` (with leading whitespace) which does not
match any valid translation.

This patch modifies the `split()` method to trim the title component.

It also refactors the SelectDatasets module to move the configuration
processing into a utility method, and allows it to be unit tested.

Additionally any white-spare around the `=` or `;` leads to erroneous
values, for example a value of:

```
Paragraph = p ;
```

Leads to:
title: `Paragraph `
value: ` p `

And multiple contiguous spaces leads to additional values, for example:
```
12pt  14pt 16pt  18pt
```

Leads to a set of values:
- `12pt`
- `` (empty value)
- `14pt`
- `16pt`
- `18pt`