Improved handling of linked images with captions
The previous code assumed <img> was a direct descendent of <figure>, so linking the image would make the editor lose sight of the caption (and re-ticking the checkbox would add a second caption).
If the image was linked before adding the caption it would be wrapped in a <figure> and would leave an empty <a> behind on the page, too.
The updated code uses a getFigure() function that searches for a parent figure to use (can be any ancestor of the <img>, not just the immediate parent).
To improve moving the image in or out of the figure a getInlineContainer() function has also been added that searches for the topmost non-block-level element of the <img> and moves that instead of the <img> itself.