Move caret behind newly inserted/updated links
Fixes #6711.

This allows the user to continue typing after inserting/updating a link, without having to fiddle with the caret position to get it out of the <a> (which takes two presses of the right arrow key, but that's hard to guess).
This behavior matches Google Docs, for both the insertion and update cases.

This affects the specific case where a link is inserted/updated through the `link` plugin's dialog when some text was previously selected or the cursor was within an existing link. It does intentionally **not** affect:
* The `quibkcars` plugin's hovering UI, which already behaves correctly. So this change makes the behaviors consistent between these two plugins.
* The `autolink` plugin, which only creates links when the user's cursor has already moved on.
* The case of creating a new link without previously selected text using the `link` plugin, i.e. where the link's text is also entered through the dialog. The plugin already behaved correctly in this case, moving the cursor after the newly inserted link.