Remove the TinyComments hack from SelectedTextLinkTest
Fixes #10143.

It's unclear to me what this comment was meant for, but it looks like it's supposed to trick the SUT into thinking that the TinyComments plugin is also active, so that the SUT would change its behavior accordingly. But the SUT does not check for the presence of TinyComments (anymore?), so this seems redundant. Additionally it's undesirable, because this HTML comment accidentally becomes the text of inserted links (rather than the URL being copied into the link text field).