TINY-12507: introduce reset layer (#10464)
Related Ticket: TINY-12507

Description of Changes:
Added CSS layer to the reset styles by wrapping all reset styles in `@layer reset {}`. This will guarantee that the reset is applied before all of the other, unlayered styles and will not fight for specificity with them. 

Pre-checks:
* [x] ~~Changelog entry added~~
* [x] ~~Tests have been added (if applicable)~~
* [x] Branch prefixed with `feature/`, `hotfix/` or `spike/`

Review:
* [x] Milestone set
* [x] ~~Docs ticket created (if applicable)~~

GitHub issues (if applicable):

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **Style**
  * Enhanced CSS structure by grouping `.tox` reset styles into a dedicated layer for better style management without affecting the user interface.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->