Add support for a CSP nonce to script and style loading.
This adds an option to add a nonce to the places where a script/ link/ style tag is created.

The nonce is mostly needed to support loading styles using `mceDefaultStyles`, in `DOMUtils.addStyle`. As well as `InitContentBody.appendStyle`, and `StyleSheetLoader.loadRawCss`. But if for whatever reason someone wants to also load scripts and styles from external resources using a nonce instead of allowed hosts, then this is now also possible.

In the editor config you can decide to set a global nonce or a specific nonce for script loading and another one for style loading.

This allows a user to remove `'unsafe-inline'` from the `style-src` CSP policy and only keep it for the `style-src-attr` policy.