A csp nonce cannot be used for files loaded from external sources.
Remove adding the CSP nonce when loading an external script (with a src attribute) or style sheet (using a link attribute).

This also removed the specific nonce for loading scripts since it was no longer used. This means that we also no longer need the `SpecificCspNonce` object since nonces only apply to `script` and `style` tags. See the spec for more info: https://www.w3.org/TR/CSP3/#match-element-to-source-list